update student ids

+++ flask_backend/models/student_enrollment.py
@@ -13,7 +13,7 @@ from models.student import Student, StudentSchema
 class StudentEnrollment(db.Model):
     id: Mapped[int] = db.Column(db.Integer, primary_key=True)
     course_number = db.Column(db.Integer, db.ForeignKey("course.course_number"), nullable=False)
-    student_id = db.Column(db.Integer, db.ForeignKey("student.student_id"), nullable=False)
+    student_id = db.Column(db.BigInteger, db.ForeignKey("student.student_id"), nullable=False)
